+++ asset-clustering/training-tuning-inference/inference.py
@@ -3,14 +3,17 @@ import logging
 import pickle
 import click
 from utils import compute_chars
-from train import normalize_host_data
+from utils import normalize_host_data
 
 
 @click.command()
 @click.option('--model', default='dbscan', help='Clustering method to use.'\
-     ' Valid choices are \'kmeans\' or \'dbscan\'. Default is \'dbscan\'')
+    ' Valid choices are \'kmeans\' or \'dbscan\'. Default is \'dbscan\'.'\
+    'The corresponding model pickle file will be read from the relative'\
+    'path \'../models/ \'.')
 @click.option('--data_fname', default='host_agg_data_day-11_day-15.csv',\
-     help='Name of the Preprocessed csv dataset to perofrm inference.')
+    help='Name of the Preprocessed csv dataset to perofrm inference. The given'\
+    'file name will be read from the relative path \'../datasets/ \'')
 @click.option('--num_days', default=5.0, help='Number of days worth of data used'\
     'in preparing the dataset. Used to normalize the features.')
 @click.option('--compute_cluster_chars', is_flag=True, help='Boolean flag. If '\
